requirements.txt
openpyxl==3.1.5
python-dotenv==1.0.1
pillow==11.0.0